更新neutron-net 至 v2.0

+++ platform/MCU/ESP8266/sdk/include/espconn.h
@@ -1,3 +1,27 @@
+/*
+ * ESPRSSIF MIT License
+ *
+ * Copyright (c) 2016 <ESPRESSIF SYSTEMS (SHANGHAI) PTE LTD>
+ *
+ * Permission is hereby granted for use on ESPRESSIF SYSTEMS ESP8266 only, in which case,
+ * it is free of charge, to any person obtaining a copy of this software and associated
+ * documentation files (the "Software"), to deal in the Software without restriction, including
+ * without limitation the rights to use, copy, modify, merge, publish, distribute, sublicense,
+ * and/or sell copies of the Software, and to permit persons to whom the Software is furnished
+ * to do so, subject to the following conditions:
+ *
+ * The above copyright notice and this permission notice shall be included in all copies or
+ * substantial portions of the Software.
+ *
+ * THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
+ * IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY, FITNESS
+ * FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE AUTHORS OR
+ * COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER LIABILITY, WHETHER
+ * IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM, OUT OF OR IN
+ * CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
+ *
+ */
+
 #ifndef __ESPCONN_H__
 #define __ESPCONN_H__
 
@@ -512,7 +536,7 @@ sint16 espconn_secure_get_size(uint8 level);
  * Returns      : result true or false
 *******************************************************************************/
 
-bool espconn_secure_ca_enable(uint8 level, uint8 flash_sector );
+bool espconn_secure_ca_enable(uint8 level, uint32 flash_sector );
 
 /******************************************************************************
  * FunctionName : espconn_secure_ca_disable
@@ -535,7 +559,7 @@ bool espconn_secure_ca_disable(uint8 level);
  * Returns      : result true or false
 *******************************************************************************/
 
-bool espconn_secure_cert_req_enable(uint8 level, uint8 flash_sector );
+bool espconn_secure_cert_req_enable(uint8 level, uint32 flash_sector );
 
 /******************************************************************************
  * FunctionName : espconn_secure_ca_disable
